feat: make use of toml.SSize where appropriate
Update Config structs to make use of toml.SSize where appropriate.
Updated fields:
- [collectd] read-buffer
- [data] tar-stream-buffer-size
- [flux-controller] query-initial-memory-bytes
- [flux-controller] query-max-memory-bytes
- [flux-controller] total-max-memory-bytes
- [http] max-body-size
- [subscriber] total-buffer-bytes
- [udp] read-buffer

+++ tsdb/engine/tsm1/engine.go
@@ -289,7 +289,7 @@ func NewEngine(id uint64, idx tsdb.Index, path string, walPath string, sfile *ts
 		Scheduler:                     newScheduler(stats, opt.CompactionLimiter.Capacity()),
 		seriesIDSets:                  opt.SeriesIDSets,
 
-		TarStreamBufferSize: opt.Config.TarStreamBufferSize,
+		TarStreamBufferSize: uint64(opt.Config.TarStreamBufferSize),
 	}
 
 	// Feature flag to enable per-series type checking, by default this is off and
